requirements.txt
streamlit==1.54.0
pandas==2.3.3
requests==2.32.5
msal==1.28.0
openpyxl==3.1.5
numpy==2.4.2
pytz==2025.2
python-dotenv==1.0.1